modules/event-consumer-service/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
pyrsistent==0.16.0